rename wood.ts -> wood-builder.ts
